feat: :rocket: Project Deployment
feat: :sparkles: Backend testing and database connection, plus application on the frontend.

The connection to the SQLite database was made, performing the basic CRUD queries and their visualization in the frontend.

feat: :construction: Implementation of Put requests and style changes.

The “Put” update request was designed within CRUD and design modifications were made.

feat: :construction: Modifications in the database and implementation of CRUD requests in the frontend.

The “type” data type was implemented for the differentiation of projects and tasks, additionally. The use of CRUD requests was implemented in the frontend for its use from the Gantt Chart.

feat: :building_construction: Code structure following the MVC architecture.

The code was reorganized to divide it and make it more flexible in its management, following the MVC architecture..

feat: :sparkles: Deployment of the database in Ralway.

The backend of the project was deployed on the railway platform for visualization.

chore: :hammer: Implementation of environment variables.

Configuration of environment variables for database management.

docs: :memo: Project Documentation

Documentation of the project explaining the tools and development.

docs: :memo: Documentation Update.

refactor: :zap: Implemented .env done

docs: :memo: Documentation fix

feat: :rocket: Deplyment Features

feat: :rocket: Deployment Stuff FIx.